requirements.txt
attrs=18.2.0=py_0
augmentor=0.2.2=pypi_0
backcall=0.1.0=py_0
blas=1.0=mkl
bleach=3.1.0=py_0
ca-certificates=2018.11.29=ha4d7672_0
cairo=1.14.12=h8948797_3
certifi=2018.11.29=py36_1000
cffi=1.11.5=py36he75722e_1
chardet=3.0.4=pypi_0
cloudpickle=0.5.3=pypi_0
cuda90=1.0=h6433d27_0
cuda91=1.0=h4c16780_0
cycler=0.10.0=pypi_0
cython=0.29.1=pypi_0
dask=0.18.1=pypi_0
decorator=4.3.0=pypi_0
dominate=2.3.1=pypi_0
entrypoints=0.3=py36_1000
expat=2.2.5=he0dffb1_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_0
fribidi=1.0.4=h14c3975_0
future=0.16.0=pypi_0
glib=2.56.1=h000015b_0
graphite2=1.3.11=h16798f4_2
graphviz=2.40.1=h21bd128_2
h5py=2.9.0=pypi_0
harfbuzz=1.8.4=hec2c2bc_0
icu=58.2=h9c2bf20_1
idna=2.7=pypi_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h24bf2e0_1002
ipython=7.3.0=py36h24bf2e0_0
ipython_genutils=0.2.0=py_1
jedi=0.13.3=py36_0
jinja2=2.10=py_1
jpeg=9b=h024ee3a_2
jsonschema=3.0.0=py36_0
jupyter_client=5.2.4=py_1
jupyter_core=4.4.0=py_0
jupyterlab=0.35.4=py36_0
jupyterlab_server=0.2.0=py_0
kiwisolver=1.0.1=pypi_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.3.0=hdf63c60_0
libgfortran-ng=7.2.0=hdf63c60_3
libpng=1.6.34=hb9fc6fc_0
libprotobuf=3.6.1=hdbcaa40_1000
libsodium=1.0.16=h14c3975_1001
libstdcxx-ng=7.3.0=hdf63c60_0
libtiff=4.0.9=he85c1e1_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markupsafe=1.1.1=py36h14c3975_0
matplotlib=2.2.2=pypi_0
medpy=0.4.0=pypi_0
mistune=0.8.4=py36h14c3975_1000
mkl=2019.1=144
mkl_fft=1.0.10=py36h14c3975_1
mkl_random=1.0.2=py36h637b7d7_2
nbconvert=5.3.1=py_1
nbformat=4.4.0=py_1
ncurses=6.1=hf484d3e_0
networkx=2.1=pypi_0
ninja=1.8.2=py36h6bb024c_1
notebook=5.7.4=py36_1000
numpy=1.14.5=pypi_0
numpy-base=1.15.4=py36hde5b4d6_0
opencv-python=3.4.1.15=pypi_0
openssl=1.1.1b=h14c3975_0
pandas=0.24.0=pypi_0
pandoc=2.6=1
pandocfilters=1.4.2=py_1
pango=1.42.2=h8589676_0
parso=0.3.4=py_0
pcre=8.42=h439df22_0
pexpect=4.6.0=py36_1000
pickleshare=0.7.5=py36_1000
pillow=5.1.0=pypi_0
pip=18.0=pypi_0
pixman=0.34.0=hceecf20_3
prometheus_client=0.6.0=py_0
prompt_toolkit=2.0.9=py_0
protobuf=3.6.1=py36hf484d3e_1001
ptyprocess=0.6.0=py36_1000
pycparser=2.19=py36_0
pydensecrf=1.0rc3=py36hf8a1672_1
pygments=2.3.1=py_0
pyparsing=2.2.0=pypi_0
pyrsistent=0.14.11=py36h14c3975_0
python=3.6.8=h0371630_0
python-dateutil=2.7.3=pypi_0
pytorch=0.4.1=py36_py35_py27__9.0.176_7.1.2_2
pytz=2018.4=pypi_0
pywavelets=0.5.2=pypi_0
pyzmq=17.0.0=pypi_0
readline=7.0=ha6073c6_4
requests=2.19.1=pypi_0
runstats=1.7.1=pypi_0
scikit-image=0.14.0=pypi_0
scikit-learn=0.19.1=pypi_0
scipy=1.1.0=pypi_0
send2trash=1.5.0=py_0
setuptools=39.2.0=py36_0
simpleitk=1.2.0=pypi_0
six=1.11.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.26.0=h67949de_1000
tensorboardx=1.6=py_0
terminado=0.8.1=py36_1001
testpath=0.4.2=py36_1000
tk=8.6.9=h84994c4_1000
toolz=0.9.0=pypi_0
torch=0.4.0=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.2.1=pypi_0
tornado=5.0.2=pypi_0
tqdm=4.23.4=pypi_0
traitlets=4.3.2=py36_1000
urllib3=1.23=pypi_0
visdom=0.1.8.3=pypi_0
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
websocket-client=0.48.0=pypi_0
wheel=0.31.1=py36_0
xz=5.2.4=h14c3975_4
zeromq=4.2.5=hf484d3e_1006
zlib=1.2.11=ha838bed_2
